+++ modules/articoli/bulk.php
@@ -19,6 +19,11 @@
 
 include_once __DIR__.'/../../core.php';
 use Modules\Articoli\Articolo;
+use Modules\Preventivi\Preventivo;
+use Modules\Anagrafiche\Anagrafica;
+use Modules\Preventivi\Components\Articolo as ArticoloPreventivo;
+use Modules\TipiIntervento\Tipo as TipoSessione;
+use Modules;
 use Prints;
 
 switch (post('op')) {
@@ -63,6 +68,71 @@ switch (post('op')) {
         exit();
 
         break;
+
+    case 'change-qta':
+        $descrizione = post('descrizione');
+        $data = post('data');
+        $qta = post('qta');
+        $n_articoli = 0;
+        
+        foreach ($id_records as $id) { 
+            $articolo = Articolo::find($id);
+            $qta_movimento = $qta - $articolo->qta;
+            $articolo->movimenta($qta_movimento, $descrizione, $data, true);
+
+            ++$n_articoli;
+        }
+
+        if ($n_articoli > 0) {
+            flash()->info(tr('Quantità cambiate a _NUM_ articoli!', [
+                '_NUM_' => $n_articoli,
+            ]));
+        } else {
+            flash()->warning(tr('Nessun articolo modificato!'));
+        }
+
+        break;
+
+    case 'crea-preventivo':
+        $nome = post('nome');
+        $data = post('data');
+        $id_tipo = post('id_tipo');
+        $id_cliente = post('id_cliente');
+        $anagrafica = Anagrafica::find($id_cliente);
+        $tipo = TipoSessione::find($id_tipo);
+        $n_articoli = 0;
+
+        $preventivo = Preventivo::build($anagrafica, $tipo, $nome, $data, 0);
+        $id_preventivo = $preventivo->id;
+        
+        foreach ($id_records as $id) { 
+            $originale = Articolo::find($id);
+            $articolo = ArticoloPreventivo::build($preventivo, $originale);
+            $articolo->qta = 1;
+            $articolo->descrizione = $originale->descrizione;
+            $articolo->um = $originale->um ?: null;
+            $articolo->costo_unitario = $originale->prezzo_acquisto;
+            $articolo->prezzo_unitario = $originale->prezzo_vendita;
+            $articolo->idiva = $originale->idiva_vendita;
+            $articolo->setPrezzoUnitario($originale->prezzo_vendita, $originale->idiva_vendita);
+            $articolo->save();
+
+            ++$n_articoli;
+        }
+
+        if ($n_articoli > 0) {
+            flash()->info(tr('_NUM_ articoli sono stati aggiunti al preventivo', [
+                '_NUM_' => $n_articoli,
+            ]));
+        } else {
+            flash()->warning(tr('Nessun articolo modificato!'));
+        }
+
+        $database->commitTransaction();
+        redirect( base_path().'/editor.php?id_module='.Modules::get('Preventivi')['id'].'&id_record='.$id_preventivo);
+        exit();
+
+        break;
 }
 
 if (App::debug()) {
@@ -95,6 +165,33 @@ $operations['stampa-etichette'] = [
         'button' => tr('Procedi'),
         'class' => 'btn btn-lg btn-warning',
         'blank' => true,
+    ],
+];
+
+$operations['change-qta'] = [
+    'text' => '<span><i class="fa fa-refresh"></i> '.tr('Aggiorna quantità').'</span>',
+    'data' => [
+        'title' => tr('Cambiare le quantità?'),
+        'msg' => tr('Per ciascun articolo selezionato, verrà modificata la quantità').'
+        <br><br>{[ "type": "text", "label": "'.tr('Quantità').'", "name": "qta", "required": 1 ]}
+        {[ "type": "text", "label": "'.tr('Causale').'", "name": "descrizione", "required": 1 ]}
+        {[ "type": "date", "label": "'.tr('Data').'", "name": "data", "required": 1, "value": "-now-" ]}',
+        'button' => tr('Procedi'),
+        'class' => 'btn btn-lg btn-warning',
+    ],
+];
+
+$operations['crea-preventivo'] = [
+    'text' => '<span><i class="fa fa-plus"></i> '.tr('Crea preventivo').'</span>',
+    'data' => [
+        'title' => tr('Creare preventivo?'),
+        'msg' => tr('Ogni articolo selezionato, verrà aggiunto al preventivo').'
+        <br><br>{[ "type": "text", "label": "'.tr('Nome preventivo').'", "name": "nome", "required": 1 ]}
+        {[ "type": "select", "label": "'.tr('Cliente').'", "name": "id_cliente", "ajax-source": "clienti", "required": 1 ]}
+        {[ "type": "select", "label": "'.tr('Tipo di attività').'", "name": "id_tipo", "values": "query=SELECT idtipointervento AS id, descrizione FROM in_tipiintervento", "required": 1 ]}
+        {[ "type": "date", "label": "'.tr('Data').'", "name": "data", "required": 1, "value": "-now-" ]}',
+        'button' => tr('Procedi'),
+        'class' => 'btn btn-lg btn-warning',
     ],
 ];