+++ modules/preventivi/edit.php
@@ -149,9 +149,21 @@ echo '
 				</div>
 			</div>
 
+            <div class="row">
+				<div class="col-md-12">
+					{[ "type": "textarea", "label": "<?php echo tr('Garanzia'); ?>", "name": "garanzia", "class": "autosize", "value": "$garanzia$" ]}
+				</div>
+			</div>
+
 			<div class="row">
 				<div class="col-md-12">
 					{[ "type": "textarea", "label": "<?php echo tr('Descrizione'); ?>", "name": "descrizione", "class": "autosize", "value": "$descrizione$" ]}
+				</div>
+			</div>
+
+            <div class="row">
+				<div class="col-md-12">
+					{[ "type": "ckeditor", "label": "<?php echo tr('Condizioni generali di fornitura'); ?>", "name": "condizioni_fornitura", "class": "autosize", "value": "$condizioni_fornitura$" ]}
 				</div>
 			</div>