+++ plugins/importFE/actions.php
@@ -129,6 +129,7 @@ switch (filter('op')) {
             'id_riga_riferimento' => post('id_riga_riferimento'),
             'movimentazione' => post('movimentazione'),
             'crea_articoli' => post('crea_articoli'),
+            'is_ritenuta_pagata' => post('is_ritenuta_pagata'),
         ];
 
         $fattura_pa = FatturaElettronica::manage($filename);
@@ -137,7 +138,7 @@ switch (filter('op')) {
         ricalcola_costiagg_fattura($id_fattura);
         elimina_scadenze($id_fattura);
         elimina_movimenti($id_fattura, 0);
-        aggiungi_scadenza($id_fattura);
+        aggiungi_scadenza($id_fattura, post('pagamento'));
         aggiungi_movimento($id_fattura, 'uscita');
 
         $fattura_pa->delete();