+++ src/API/App/v1/Clienti.php
@@ -21,6 +21,7 @@ namespace API\App\v1;
 
 use API\App\AppResource;
 use Auth;
+use Modules\Anagrafiche\Anagrafica;
 
 class Clienti extends AppResource
 {
@@ -78,6 +79,7 @@ class Clienti extends AppResource
         // Gestione della visualizzazione dei dettagli del record
         $query = 'SELECT an_anagrafiche.idanagrafica AS id,
             an_anagrafiche.ragione_sociale,
+            an_anagrafiche.tipo,
             an_anagrafiche.piva AS partita_iva,
             an_anagrafiche.codice_fiscale,
             an_anagrafiche.indirizzo,
@@ -104,4 +106,62 @@ class Clienti extends AppResource
 
         return $record;
     }
+
+    public function createRecord($data)
+    {
+
+        $ragione_sociale = $data['ragione_sociale'];
+        $id_tipo = [1];
+
+        $anagrafica = Anagrafica::build($ragione_sociale, null, null, $id_tipo);
+        $id_record = $anagrafica->id;
+
+        $anagrafica->ragione_sociale = $data['ragione_sociale'];
+        $anagrafica->tipo = $data['tipo'];
+        $anagrafica->piva = $data['partita_iva'];
+        $anagrafica->codice_fiscale = $data['codice_fiscale'];
+        $anagrafica->indirizzo = $data['indirizzo'];
+        $anagrafica->cap = $data['cap'];
+        $anagrafica->citta = $data['citta'];
+        $anagrafica->provincia = $data['provincia'];
+        $anagrafica->telefono = $data['telefono'];
+        $anagrafica->cellulare = $data['cellulare'];
+        $anagrafica->email = $data['email'];
+
+        $anagrafica->save();
+
+        return [
+            'id' => $id_record,
+        ];
+
+    }
+
+    public function updateRecord($data)
+    {
+        $anagrafica = Anagrafica::find($data['id']);
+
+        $this->aggiornaRecord($anagrafica, $data);
+        $anagrafica->save();
+
+        return [];
+    }
+
+    protected function aggiornaRecord($record, $data)
+    {
+        $database = database();
+
+        // Aggiornamento anagrafica
+        $record->ragione_sociale = $data['ragione_sociale'];
+        $record->tipo = $data['tipo'];
+        $record->piva = $data['partita_iva'];
+        $record->codice_fiscale = $data['codice_fiscale'];
+        $record->indirizzo = $data['indirizzo'];
+        $record->cap = $data['cap'];
+        $record->citta = $data['citta'];
+        $record->provincia = $data['provincia'];
+        $record->telefono = $data['telefono'];
+        $record->cellulare = $data['cellulare'];
+        $record->email = $data['email'];
+    }
+
 }