+++ templates/preventivi/body.php
@@ -386,8 +386,22 @@ echo '
             '.nl2br($documento['esclusioni']).'
         </td>
     </tr>
+
+    <tr>
+        <th>
+            '.tr('Garanzia', [], ['upper' => true]).'
+        </th>
+
+        <td>
+            '.nl2br($documento['garanzia']).'
+        </td>
+    </tr>
 </table>';
 
 // Conclusione
 echo '
 <p class="text-center">'.tr("In attesa di un Vostro Cortese riscontro, colgo l'occasione per porgere Cordiali Saluti").'</p>';
+
+if(!empty($documento->condizioni_fornitura)){
+    echo '<pagebreak>'.$documento->condizioni_fornitura;
+}